+++ .github/workflows/ci.yml
@@ -332,6 +332,7 @@ jobs:
           Avatars,
           Console,
           Databases,
+          TablesDB,
           Functions,
           FunctionsSchedule,
           GraphQL,
@@ -420,7 +421,7 @@ jobs:
             # Services that rely on sequential test method execution (shared static state)
             FUNCTIONAL_FLAG="--functional"
             case "${{ matrix.service }}" in
-              Databases|Functions|Realtime) FUNCTIONAL_FLAG="" ;;
+              Databases|TablesDB|Functions|Realtime) FUNCTIONAL_FLAG="" ;;
             esac
 
             docker compose exec -T \
