+++ tests/e2e/Services/TablesDB/TablesDBCustomClientTest.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Tests\E2E\Services\Databases;
+namespace Tests\E2E\Services\TablesDB;
 
+use Tests\E2E\Services\Databases\DatabasesBase;
 use Tests\E2E\Scopes\ApiTablesDB;
 use Tests\E2E\Scopes\ProjectCustom;
 use Tests\E2E\Scopes\Scope;
