+++ tests/e2e/Services/TablesDB/Transactions/TablesDBACIDTest.php
@@ -1,18 +1,19 @@
 <?php
 
-namespace Tests\E2E\Services\Databases\Transactions;
+namespace Tests\E2E\Services\TablesDB\Transactions;
 
+use Tests\E2E\Services\Databases\Transactions\ACIDBase;
 use Tests\E2E\Scopes\ApiTablesDB;
 use Tests\E2E\Scopes\ProjectCustom;
 use Tests\E2E\Scopes\Scope;
-use Tests\E2E\Scopes\SideServer;
+use Tests\E2E\Scopes\SideClient;
 use Tests\E2E\Traits\DatabasesUrlHelpers;
 
-class TablesDBTransactionsCustomServerTest extends Scope
+class TablesDBACIDTest extends Scope
 {
-    use TransactionsBase;
+    use ACIDBase;
     use DatabasesUrlHelpers;
     use ProjectCustom;
-    use SideServer;
+    use SideClient;
     use ApiTablesDB;
 }
