+++ tests/e2e/Services/TablesDB/Transactions/TablesDBTransactionPermissionsCustomClientTest.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace Tests\E2E\Services\Databases\Transactions;
+namespace Tests\E2E\Services\TablesDB\Transactions;
 
+use Tests\E2E\Services\Databases\Transactions\TransactionPermissionsBase;
 use Tests\E2E\Scopes\ApiTablesDB;
 use Tests\E2E\Scopes\ProjectCustom;
 use Tests\E2E\Scopes\Scope;
