+++ tests/e2e/Services/TablesDB/Transactions/TablesDBTransactionsCustomServerTest.php
@@ -1,18 +1,19 @@
 <?php
 
-namespace Tests\E2E\Services\Databases\Transactions;
+namespace Tests\E2E\Services\TablesDB\Transactions;
 
+use Tests\E2E\Services\Databases\Transactions\TransactionsBase;
 use Tests\E2E\Scopes\ApiTablesDB;
 use Tests\E2E\Scopes\ProjectCustom;
 use Tests\E2E\Scopes\Scope;
-use Tests\E2E\Scopes\SideConsole;
+use Tests\E2E\Scopes\SideServer;
 use Tests\E2E\Traits\DatabasesUrlHelpers;
 
-class TablesDBTransactionsConsoleClientTest extends Scope
+class TablesDBTransactionsCustomServerTest extends Scope
 {
     use TransactionsBase;
     use DatabasesUrlHelpers;
     use ProjectCustom;
-    use SideConsole;
+    use SideServer;
     use ApiTablesDB;
 }
